requirements.txt
streamlit>=1.0.0
Pillow>=9.0.0  
keras
opencv-python
numpy>=1.18.0
pandas>=1.2.0
matplotlib>=3.4.0
